+++ source/io.c
@@ -28,7 +28,7 @@ jlong JNICALL Java_software_amazon_awssdk_crt_EventLoopGroup_event_1loop_1group_
 
     struct aws_allocator *allocator = aws_jni_get_allocator();
     struct aws_event_loop_group *elg = aws_mem_acquire(allocator, sizeof(struct aws_event_loop_group));
-    int result = aws_event_loop_group_default_init(elg, allocator, num_threads);
+    int result = aws_event_loop_group_default_init(elg, allocator, (uint16_t)num_threads);
     if (result != AWS_OP_SUCCESS) {
         aws_jni_throw_runtime_exception(
             env, "EventLoopGroup.event_loop_group_new: aws_event_loop_group_default_init failed");
